+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- [server] Add methods for manual client management
+
 ## 0.20.1 - 2018-04-23
 
 - [server] Add a workaround in `Resource` definition to avoid https://github.com/rust-lang/rust/issues/50153